requirements.txt
wandb>=0.15.1
numpy>=1.24.2
Pillow>=9.5.0
mpi4py-mpich>=3.1.2
tensorboard>=2.12.2
packaging
einops
torchinfo